requirements.txt
alembic==1.13.1
asyncpg==0.30.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-qt5==0.1.10
pytokens==0.2.0
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.29
tqdm==4.67.1
types-pytz==2025.2.0.20250809
typing_extensions==4.15.0
tzdata==2025.2
uliweb-alembic==0.6.9
urllib3==2.5.0
